refactor(Car): 정적 팩토리 메소드로 이전
